Basic Workflow
This Workflow explains what is the use of the programming schedule is to send an automatically email when the assistant adds one or more leads with conditional IF. It is the same exercise for the supplier when the time is completed. The most important thing is when the assistant can JSON to personalize the email; it is amazing.