Form exercise
This workflow helps the assistant to get data about leads when registering personal information on the form and then update a row in the Google Sheet database and finish with sending the email to the assistant with relevant information about the clients' prospects. The objective is to close the sale.