¿Cripto or Bitcoin?
This is the normal inversion of the person when talking or sharing in Google regarding ¿Cripto or Bitcoin?

The objective of this workflow is to help less experienced team members understand the risks associated with cryptocurrency investments by utilizing a public document from Libre University in Colombia with a database in OpenIA to answer the prevention.
**Guido Rivas**